transfer burger menu to navbar, fix layout
make leave button reusable

#191
