+++ app/components/Table.tsx
@@ -7,7 +7,6 @@ import { AppContext } from '@/app/contexts/AppStoreProvider';
 import { sendLog } from '../hooks/server_actions';
 import { SocketContext } from '../contexts/WebSocketProvider';
 import Felt from './Felt';
-import TableMenuBurger from './TableMenuBurger/TableMenuBurger';
 
 const initialPlayers: (Player | null)[] = [
     null,
@@ -238,7 +237,6 @@ const Table = () => {
             maxWidth="100%"
             overflow="hidden"
         >
-            <TableMenuBurger />
             <Box
                 as="picture"
                 className="table-image"
